refactor(gateway): GET /v1/models sai para openai_models.rs
O #1029 levou openai_api.rs de 1346 para 1523 linhas e o ratchet acusou
a passagem do limite de 1500 (`files_over_1500` 21 -> 22). O endpoint de
modelos e uma responsabilidade propria — a regra que decide o que o
gateway serve espelha o roteamento do campo `model`, e testa-la nao
precisa de nada do caminho de chat —, entao vai para `openai_models`
com os dois testes. openai_api.rs volta a 1306 linhas, abaixo do que
era antes do #1029. As duas rotas (`router.rs` e `build_openai_router`)
apontam para o modulo novo; nada de comportamento muda.
